refactor: Relocate GBKit settings logic to hook
Reduce the scope of concerns for the visual editor component.
