+++ src/components/visual-editor/index.jsx
@@ -6,7 +6,7 @@ import clsx from 'clsx';
 /**
  * WordPress dependencies
  */
-import { useRef, useMemo } from '@wordpress/element';
+import { useRef } from '@wordpress/element';
 import {
 	BlockList,
 	privateApis as blockEditorPrivateApis,
@@ -14,13 +14,11 @@ import {
 import { Popover } from '@wordpress/components';
 import {
 	store as editorStore,
-	mediaUpload,
 	EditorSnackbars,
 	PostTitle,
-	privateApis as editorPrivateApis,
 } from '@wordpress/editor';
 import { useSelect } from '@wordpress/data';
-import { store as coreStore, useEntityBlockEditor } from '@wordpress/core-data';
+import { useEntityBlockEditor } from '@wordpress/core-data';
 // Default styles that are needed for the editor.
 import '@wordpress/components/build-style/style.css';
 import '@wordpress/block-editor/build-style/style.css';
@@ -43,12 +41,12 @@ import { unlock } from '../../lock-unlock';
 import { useMediaUpload } from '../../hooks/use-media-upload';
 import { useHostBridge } from './use-host-bridge';
 import { useEditorSetup } from './use-editor-setup';
+import { useGBKitSettings } from './use-gbkit-settings';
 
 /**
  * @typedef {import('../utils/bridge').Post} Post
  */
 
-const { useBlockEditorSettings } = unlock(editorPrivateApis);
 const {
 	ExperimentalBlockEditorProvider: BlockEditorProvider,
 	ExperimentalBlockCanvas: BlockCanvas,
@@ -74,56 +72,13 @@ function VisualEditor({ post }) {
 		};
 	}, []);
 
-	const {
-		blockPatterns,
-		editorSettings,
-		hasUploadPermissions,
-		reusableBlocks,
-	} = useSelect(
-		(select) => {
-			const { getEntityRecord, getEntityRecords } = select(coreStore);
-			const { getEditorSettings } = select(editorStore);
-			const user = getEntityRecord('root', 'user', post.author);
-
-			return {
-				editorSettings: getEditorSettings(),
-				blockPatterns: select(coreStore).getBlockPatterns(),
-				hasUploadPermissions: user?.capabilities?.upload_files ?? true,
-				reusableBlocks: getEntityRecords('postType', 'wp_block'),
-			};
-		},
-		[post.author]
-	);
-
 	const [postBlocks, onInput, onChange] = useEntityBlockEditor(
 		'postType',
 		post.type,
 		{ id: post.id }
 	);
 
-	const blockEditorSettings = useBlockEditorSettings(
-		editorSettings,
-		post.type,
-		post.id,
-		'visual'
-	);
-
-	const settings = useMemo(
-		() => ({
-			...blockEditorSettings,
-			hasFixedToolbar: true,
-			mediaUpload: hasUploadPermissions ? mediaUpload : undefined,
-			__experimentalReusableBlocks: reusableBlocks,
-			__experimentalBlockPatterns: blockPatterns,
-		}),
-		[
-			blockEditorSettings,
-			blockPatterns,
-			hasUploadPermissions,
-			reusableBlocks,
-		]
-	);
-
+	const settings = useGBKitSettings(post);
 	const styles = useEditorStyles();
 	useMediaUpload();
